backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
pillow==11.3.0
scikit-image==0.25.2
numpy==2.3.2
httpx==0.28.1
pytest==8.4.1
